Persistence implementation splitted into 2 different classes.
