+++ apg-patch-service-api/src/main/java/com/apgsga/microservice/patch/api/PatchPersistence.java
@@ -35,5 +35,4 @@ public interface PatchPersistence {
 
 	void clean();
 
-	void init() throws IOException;
 }
