+++ apg-patch-service-core/src/main/java/com/apgsga/microservice/patch/core/impl/persistence/utils/DbModulesUtil.java
@@ -1,5 +1,6 @@
 package com.apgsga.microservice.patch.core.impl.persistence.utils;
 
+import java.io.IOException;
 import java.util.List;
 
 import org.springframework.core.io.FileSystemResourceLoader;
@@ -15,7 +16,7 @@ public class DbModulesUtil {
 
 	private static List<String> dbModulesList = Lists.newArrayList("testdbmodule", "testdbAnotherdbModule");
 
-	public static void main(String[] args) {
+	public static void main(String[] args) throws IOException {
 		final ResourceLoader rl = new FileSystemResourceLoader();
 		final PatchPersistence db = new FilebasedPatchPersistence(rl.getResource("db"),rl.getResource("work"));
 		DbModules intialLoad = new DbModules(dbModulesList);
